Code quality: Fixed redundant type cast.

+++ src/org/parabuild/ci/webui/admin/AbstractLogConfigPanel.java
@@ -95,7 +95,7 @@ public abstract class AbstractLogConfigPanel extends MessagePanel implements Val
    * @param type log type to set.
    */
   public final void setLogType(final byte type) {
-    flLogType.setCode((int) type);
+    flLogType.setCode(type);
   }
